langchain-text-splitters/requirements.txt
langchain-experimental
langchain-google-genai
python-dotenv
google-generativeai
